requirements.txt
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
bokeh==0.11.1
certifi==2016.2.28
cffi==1.6.0
cryptography==1.3.2
enum34==1.1.6
Flask==0.10.1
futures==3.0.5
gunicorn==19.6.0
idna==2.1
inflection==0.3.1
ipaddress==1.0.16
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
more-itertools==2.2
ndg-httpsclient==0.4.0
numpy==1.11.0
pandas==0.18.1
pyasn1==0.1.9
pycparser==2.14
pyOpenSSL==16.0.0
python-dateutil==2.5.3
pytz==2016.4
PyYAML==3.11
Quandl==3.0.1
requests==2.10.0
singledispatch==3.4.0.3
six==1.10.0
tornado==4.3
Werkzeug==0.11.9
